requirements.txt
flask==2.0.1
google-cloud-vision==2.4.2
python-dotenv==0.19.0
Werkzeug==2.0.1
